chore: onboarding 데이터 상수화
