+++ src/constants/onboarding.ts
@@ -1,0 +1,27 @@
+export const OnboardingData = [
+  {
+    id: 1,
+    title: "원하는 제품을 골라요",
+    image: "/images/onboarding/onboarding-1.png",
+  },
+  {
+    id: 2,
+    title: "쇼핑앱에서 구매해요",
+    image: "/images/onboarding/onboarding-2.png",
+  },
+  {
+    id: 3,
+    title: "챌린저스에 올려요",
+    image: "/images/onboarding/onboarding-3.png",
+  },
+  {
+    id: 4,
+    title: "구매한 돈을 돌려받아요!",
+    image: "/images/onboarding/onboarding-4.png",
+  },
+  {
+    id: 5,
+    title: "자유롭게 출금까지!",
+    image: "/images/onboarding/onboarding-5.png",
+  },
+];
